+++ .github/generate-workflows.py
@@ -107,7 +107,7 @@ for documentclass in documentclasses:
                               yml.write("           --bibtextool=%s\\\n" % bibtextool)
                               yml.write("           --overleaf=false\\\n")
                               yml.write("           --texlive=%s\\\n" % texlive)
-                              yml.write("           --docker=reitzig\\\n")
+                              yml.write("           --docker=iot\\\n")
                               yml.write("           --lang=%s\\\n" % language)
                               yml.write("           --font=%s\\\n" % font)
                               yml.write("           --listings=%s\\\n" % listing)